requirements.txt
streamlit==1.32.0
google-generativeai==0.7.2
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.0
openpyxl==3.1.2
Pillow==10.2.0
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0